+++ src/db/alembic/versions/c8333c9b442e_create_all_tables.py
@@ -1,0 +1,34 @@
+"""create all tables
+
+Revision ID: c8333c9b442e
+Revises: 
+Create Date: 2022-10-15 23:55:46.159549
+
+"""
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import mysql
+
+# revision identifiers, used by Alembic.
+revision = 'c8333c9b442e'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('article_keywords')
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('article_keywords',
+    sa.Column('id_article', mysql.INTEGER(display_width=11), autoincrement=False, nullable=True),
+    sa.Column('id_keyword', mysql.BIGINT(display_width=20), autoincrement=False, nullable=True),
+    mariadb_collate='utf8mb4_bin',
+    mariadb_default_charset='utf8mb4',
+    mariadb_engine='InnoDB'
+    )
+    # ### end Alembic commands ###
